am261x: iar: Add IAR toolchain support with examples and libraries
Infrastructure:
- Port assembly codes, startup scripts, and linker scripts to IAR syntax
- Add IAR project templates and genProjectSpec script support
- Add syscfg viewer setup for IAR
- Add multicore project infrastructure for IAR

Library Support:
- Port FreeRTOS assembly codes to IAR syntax
- Port USB SYNP and TinyUSB libraries for IAR

Examples Ported:
- Hello World, gpio led blink, i2c led blink, usb cdc echo,
  dpl, empty, ospi flash/file io, usb dfu, usb ncm

Project Updates:
- Separate syscfg files for IAR and TI ARM Clang toolchains
- Update IAR project file paths for new folder structure
- Remove obsolete linker files
- Fix syscfg file paths for tests

Compiler Warning Fixes:
- Fix type cast warnings in drivers and kernel
- Fix volatile-related warnings in kernel
- Remove unnecessary volatiles and type casts in USB/TUSB drivers
- Suppress Pe177 warning in IAR builds
- Suppress third-party file warnings
- Fix warnings in EPWM and OSPI EDMA files

Fixes: MCUSDK-15031

+++ source/drivers/bootloader/soc/am261x/bootloader_soc.c
@@ -505,7 +505,7 @@ int32_t Bootloader_authStart(uintptr_t startAddr, uint32_t size)
 {
     int32_t status = SystemP_FAILURE;
 
-    gSecureBootStreamArray[gStreamId].dataIn = (uint8_t *)SOC_virtToPhy((void *)startAddr);
+    gSecureBootStreamArray[gStreamId].dataIn = (uint8_t *)(uintptr_t)SOC_virtToPhy((void *)startAddr);
     gSecureBootStreamArray[gStreamId].dataLen = size;
     gSecureBootStreamArray[gStreamId].canBeEncrypted = BOOTLOADER_APP_SEGMENT_CANNOTBE_ENCRYPTED;
 
@@ -520,7 +520,7 @@ int32_t Bootloader_authUpdate(uintptr_t startAddr, uint32_t size, uint8_t enc)
 {
 	int32_t status = SystemP_FAILURE;
 
-    gSecureBootStreamArray[gStreamId].dataIn = (uint8_t *)SOC_virtToPhy((void *)startAddr);
+    gSecureBootStreamArray[gStreamId].dataIn = (uint8_t *)(uintptr_t)SOC_virtToPhy((void *)startAddr);
     gSecureBootStreamArray[gStreamId].dataLen = size;
     gSecureBootStreamArray[gStreamId].canBeEncrypted = enc;
