+++ source/drivers/lin/v0/internal/sci_lin_internal.c
@@ -295,7 +295,7 @@ void LIN_HLD_setMaxBaudRate(uint32_t baseAddr, uint32_t linClk,
      * - `baudrate` is the desired communication baud rate.
     */
     HW_WR_FIELD32_RAW(  (baseAddr + CSL_LIN_MBRSR), CSL_LIN_MBRSR_MBR_MASK,
-                        CSL_LIN_MBRSR_MBR_SHIFT, (0.9 * linClk / maxBaud));
+                        CSL_LIN_MBRSR_MBR_SHIFT, (uint32_t)(0.9 * linClk / maxBaud));
 }
 
 void LIN_HLD_setFrameLength(uint32_t baseAddr, uint8_t length)
